Updated Search Parameters to take in double. Flickr API allows value of radius to be less than 1, i.e 0.1

+++ Flickr4Java/src/main/java/com/flickr4java/flickr/photos/SearchParameters.java
@@ -67,7 +67,7 @@ public class SearchParameters {
 
     private String longitude;
 
-    private int radius = -1;
+    private double radius = -1;
 
     private String radiusUnits;
 
@@ -660,7 +660,7 @@ public class SearchParameters {
         latitude = lat;
     }
 
-    public void setRadius(int r) {
+    public void setRadius(double r) {
         radius = r;
     }
